refactor main screen
