+++ apps/desktop/index.d.ts
@@ -5,7 +5,6 @@ import type { WebviewTag } from "electron";
 
 interface ImportMetaEnv {
 	readonly DEV_SERVER_PORT?: string;
-	readonly ENABLE_NEW_UI?: string;
 }
 
 interface ImportMeta {
