+++ apps/desktop/src/renderer/env.d.ts
@@ -1,7 +1,7 @@
 /// <reference types="vite/client" />
 
 interface ImportMetaEnv {
-	readonly ENABLE_NEW_UI: string;
+	// Environment variables
 }
 
 interface ImportMeta {
