+++ apps/desktop/src/renderer/screens/main/components/Avatar.tsx
