+++ apps/desktop/src/renderer/screens/main/components/StatusIndicator.tsx
