+++ apps/desktop/src/renderer/screens/main/components/TaskListItem.tsx
