+++ apps/desktop/src/renderer/screens/main/components/TaskPreview.tsx
