+++ apps/desktop/src/renderer/screens/main/components/TaskTabs/SidebarToggle.tsx
