+++ apps/desktop/src/renderer/screens/main/components/TaskTabs/TaskWorktreeContent.tsx
