+++ apps/desktop/src/renderer/screens/main/components/TaskTabs/index.ts
