+++ apps/desktop/src/renderer/screens/main/components/TaskTabs/types.ts
