+++ apps/desktop/src/renderer/screens/main/components/mock-data.ts
