+++ apps/desktop/src/renderer/screens/main/components/views/BrowserView.tsx
