+++ apps/desktop/src/renderer/screens/main/components/views/SummaryView.tsx
