+++ apps/desktop/src/renderer/screens/main/components/views/TerminalsView.tsx
